WIP: run doctests, adjust Chebyshev printing
