+++ .gitignore
@@ -5,3 +5,4 @@ docs/site/
 *.jl.mem
 
 Manifest.toml
+archive/
